fix(86byb8er1): fixed username getting hashed instead of password
