+++ src/main/java/pl/wip/wipplatformbe/requests/RegisterRequest.java
@@ -1,22 +1,7 @@
 package pl.wip.wipplatformbe.requests;
 
-public class RegisterRequest {
-    private String username;
-    private String password;
+import pl.wip.wipplatformbe.requests.base.AuthenticationRequest;
 
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
+public class RegisterRequest extends AuthenticationRequest {
+    
 }
