+++ src/main/java/pl/wip/wipplatformbe/requests/base/AuthenticationRequest.java
@@ -1,0 +1,14 @@
+package pl.wip.wipplatformbe.requests.base;
+
+public abstract class AuthenticationRequest {
+    private String username;
+    private String password;
+
+    public String getUsername() {
+        return username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+}
